Create a presentation from a targeter report object

This function generates a presentation (slides), currently targeting powerpoint (other formats to come).

Usage

slidify(
  object,
  format = "pptx",
  template = NULL,
  pptx_template = NULL,
  generate_yaml = TRUE,
  output_dir = ".",
  output_file = NULL,
  metadata = NULL,
  delete_temporary_files = TRUE,
  ...
)

Arguments

object

- an object of class targeter

format

- character, default: pptx (currently the only format provided)

template

- character. path to a Quarto qmd document that will be used to generate presentation. If NULL (default), we will use targeter package default template.

pptx_template

- charcater. Path to a default PPTX that will be used as Quarto->>PPTX generation. If NULL (default), we will use targeter package default PPTX template.

generate_yaml

- boolean, defauly TRUE. For targeter defaults template we separated YAML header as we generate programmatically it (some values in YAML depend on parameters)

output_dir

- character. Output directory - default to working directory

output_file

- character, name of output file (with extension). if NULL (default), slidify will generate a temporary name.

metadata

- not used currently

delete_temporary_files

- logical. Whether we delete temporary files or not once presentation is generated. Default: TRUE. On might want to use FALSE for debugging purpose.

Value

inviisibly returns path to the generated presentation

Examples

if (FALSE) { # \dontrun{
tar <- targeter(adult, target ="ABOVE50K", analysis_name="Analyse",
naming_conventions=FALSE)
slidfy(tar)
} # }