tar_report

This function creates an automatic report according to a predefined template in the package or a user-generated template. Create a report from a targeter tree object

Usage

# S3 method for class 'tartree'
tar_report(
  object,
  metadata = NULL,
  metadata_vars = list(varname = "variable", varlabel = "label"),
  format = "html",
  template = NULL,
  quarto_root_dir = ".",
  quarto_targeters_project_dir = "targeter-reports",
  quarto_project_template = getOption("targeter.template",
    file.path(find.package("targeter", lib.loc = .libPaths()), "ressources",
    "_targeter_project_default.yml")),
  quarto_project_brandfile = getOption("targeter.brandfile",
    file.path(find.package("targeter", lib.loc = .libPaths()), "ressources",
    "_targeter_brand_default.yml")),
  targeter_sub_folder = paste0("targeter-report-tree-", format(Sys.time(), format =
    "%Y-%m-%d_%H%M%S")),
  pptx_reference_doc = NULL,
  output_file = "index",
  title = paste(attr(object, "tar_object")$analysis, "- decision tree"),
  author = getOption("targeter.author", "welovedatascience targeter"),
  render = TRUE,
  freeze = TRUE,
  debug = FALSE,
  verbose = FALSE,
  custom_fields = list(report_type = "targeter-tree", freeze = freeze),
  ...
)

Arguments

object

- an object of class targeter tartree

metadata

- data.frame - if metadata is loaded in R environment, label of the variables can be used. Default value (NULL) corresponds to no metadata available. The label will be used for the title and the x-axis of the graph.

metadata_vars

- list of two character strings: varname and varlabel. Default to varname="variable", varlabel="label". Used to specify the columns in metadata that contains variable names and labels.

format

- Target output format (defaults to "html"). The option "all" will render all formats.

template

- character. path to a Quarto qmd document that will be used to generate presentation. If NULL (default), we will use targeter package default template.

quarto_root_dir

(default "./targeter-reports") Path to a root folder where quarto project folder will be avaible. Default to "." so that calling targeter first time in an analyses project will initiate everything for user.

quarto_targeters_project_dir

(default "targeter-reports", within quarto_root_dir) Path to a folder where all targeter reports will be stored. Using a pre-existing projetc allows benfiting from its settings. (you might want to put brand_file to empty or use your own brand file)

quarto_project_template

- character. Path to a default Quarto template that will be used as Quarto reference-doc for all formats. If NULL (default), we will use targeter package default templates.

quarto_project_brandfile

- character. Path to a default Quarto brandfile that will be used as Quarto reference-doc for all formats. If NULL (default), we will use targeter package default brandfile.

targeter_sub_folder

- character. Output directory - default to create a new dedicated directory in working folder.

output_file

- character, name of output file (without file extension). default value: index

title

- character, title for the presentation. Overide default title that consists in taking title for targeter object 'analysis' slot.

author

- character: presentatiopn author.

render

- logical. If TRUE (default), the report will be rendered and saved in the target_sub_folder. If FALSE, the report will be generated but not rendered. Useful for debugging.

freeze

- logical: default TRUE. If TRUE, the report will be frozen (no more changes can be made). If FALSE, the report will be editable. Frozen reports renders can always be bypassed by explicit call to quarto.

debug

- logical. If TRUE, the function will print debug information (passed to Quarto call).

verbose

- logical. If TRUE, the function will print some information along the process

custom_fields

- list of custom fields to be added to the report.

...

additional parameters to be passed to quarto, for instance allowing to pass/overwite any YAML set-up

Value

invisibly returns path to the generated specific file (unique format) or folder (several formats)