Cross a single variable with a target

this is a convenience function that behind the scene calls complete targeter function on data. This allows to quickly assess relation for one variable and have access to plots. It is highly recommended to have a look at targeter parameters as they could be used here also.

Usage

crossvar(data, target, var, ...)

Arguments

data

data.frame or data.table

target

character - name of the target variable of interest

var

character - name of the candidate variable to be crossed with target

...

Additional parameter to be passed to targeter

Value

returns an object of class crossvar

Details

content of returned object depends on the type of the target (binary/categorical/continuous.) For more information, look at the documentation of the crossvar class.

See also

Examples

if (FALSE) { # \dontrun{
if(interactive()){
 data(adult)
 crossvar(adult, target='ABOVE50K', var='AGE')
 t <- crossvar(adult, target='HOURSPERWEEK', var='FNLWGT')
 plot(t)
 }
} # }