Function focus()

The function subset the targeter object and retains the n top discriminant variables with regards to a criteria. It also allows filtering on minimum number of values within modality which is useful when using criteria index.max.index.

Usage

focus(
  x,
  criteria = c("IV", "index.max.index", "chisquare", "pvalue", "index.max.count",
    "index.max.props"),
  n = 10,
  nmin = NULL,
  min_criteria = NULL,
  force_vars = character(),
  summary_object = NULL
)

Arguments

x

an object of class "targeter".

criteria

character - sort the output according to (by default, order by decreasing index.max.index):

"IV" - information value.

"index.max.index" -value of the maximum index.

"chisquare" - chi-square value.

"pvalue" - p-value.

"index.max.count" - number of observations in the class associated to the maximum index.

"index.max.props" - proportion of the class associated to the maximum index.

n

integer - number of top variables to select. By default, n = 10.

nmin

integer - minimum number of profiles in the class. By default, 100.

min_criteria

- By default NULL. If it's specified, only the observations that have criteria >= min_criteria.

force_vars

- character, list of variables that will be kept and then won't never be filtered.

summary_object

- By default NULL. If provided, must be pre-computed summary on x callcross var object

Value

It returns the filtered input object.

See also

Examples

data(adult)
t <- targeter(adult, target = "ABOVE50K", analysis_name = "Analyse")
#> 
#> INFO:target ABOVE50K detected as type: binary
#> INFO:binary target contains number, automatic chosen level: 1; override using `target_reference_level`
focus(t)
#> $dataname
#> [1] "adult"
#> 
#> $target
#> [1] "ABOVE50K"
#> 
#> $description_data
#> NULL
#> 
#> $description_target
#> NULL
#> 
#> $target_type
#> [1] "binary"
#> 
#> $target_stats
#>   value count nNA percNA      perc
#> 1     0 24720   0      0 0.7591904
#> 2     1  7841   0      0 0.2408096
#> 
#> $analysis
#> [1] "Analyse"
#> 
#> $date
#> [1] "2023-09-13"
#> 
#> $profiles
#> $profiles$AGE
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: AGE of type: numeric
#> 
#>                      0    1         0            1
#> [a] from 17 to 21 2408    2 0.9991701 0.0008298755
#> [b] from 21 to 24 2334   28 0.9881456 0.0118543607
#> [c] from 24 to 28 3031  228 0.9300399 0.0699601105
#> [d] from 28 to 31 2117  424 0.8331366 0.1668634396
#> [e] from 31 to 34 2028  563 0.7827094 0.2172906214
#> [f] from 34 to 37 1937  723 0.7281955 0.2718045113
#> [g] from 37 to 40 1649  852 0.6593363 0.3406637345
#> [h] from 40 to 44 2062 1090 0.6541878 0.3458121827
#> [i] from 44 to 48 1754 1149 0.6042025 0.3957974509
#> [j] from 48 to 52 1391  926 0.6003453 0.3996547259
#> [k] from 52 to 59 1795 1071 0.6263084 0.3736915562
#> [l] from 59 to 90 2214  785 0.7382461 0.2617539180
#> 
#> Infomation value: 1.218432
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$WORKCLASS
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: WORKCLASS of type: character
#> 
#>                       0    1         0         1
#>  ?                 1645  191 0.8959695 0.1040305
#>  Federal-gov        589  371 0.6135417 0.3864583
#>  Local-gov         1476  617 0.7052078 0.2947922
#>  Never-worked         7    0 1.0000000 0.0000000
#>  Private          17733 4963 0.7813271 0.2186729
#>  Self-emp-inc       494  622 0.4426523 0.5573477
#>  Self-emp-not-inc  1817  724 0.7150728 0.2849272
#>  State-gov          945  353 0.7280431 0.2719569
#>  Without-pay         14    0 1.0000000 0.0000000
#> 
#> Infomation value: 0.166504
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$EDUCATION
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: EDUCATION of type: character
#> 
#>                  0    1         0          1
#>  10th          871   62 0.9335477 0.06645230
#>  11th         1115   60 0.9489362 0.05106383
#>  12th          400   33 0.9237875 0.07621247
#>  1st-4th       162    6 0.9642857 0.03571429
#>  5th-6th       317   16 0.9519520 0.04804805
#>  7th-8th       606   40 0.9380805 0.06191950
#>  9th           487   27 0.9474708 0.05252918
#>  Assoc-acdm    802  265 0.7516401 0.24835989
#>  Assoc-voc    1021  361 0.7387844 0.26121563
#>  Bachelors    3134 2221 0.5852474 0.41475257
#>  Doctorate     107  306 0.2590799 0.74092010
#>  HS-grad      8826 1675 0.8404914 0.15950862
#>  Masters       764  959 0.4434127 0.55658735
#>  Preschool      51    0 1.0000000 0.00000000
#>  Prof-school   153  423 0.2656250 0.73437500
#>  Some-college 5904 1387 0.8097655 0.19023454
#> 
#> Infomation value: 0.753711
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$EDUCATIONNUM
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: EDUCATIONNUM of type: numeric
#> 
#>                      0    1         0          1
#> [a] from 1 to 7   2494  151 0.9429112 0.05708885
#> [b] from 7 to 9   1515   93 0.9421642 0.05783582
#> [c] from 9 to 10  8826 1675 0.8404914 0.15950862
#> [d] from 10 to 12 6925 1748 0.7984550 0.20154502
#> [e] from 12 to 13  802  265 0.7516401 0.24835989
#> [f] from 13 to 16 4158 3909 0.5154332 0.48456675
#> 
#> Infomation value: 0.662038
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$MARITALSTATUS
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: MARITALSTATUS of type: character
#> 
#>                            0    1         0          1
#>  Divorced               3980  463 0.8957911 0.10420887
#>  Married-AF-spouse        13   10 0.5652174 0.43478261
#>  Married-civ-spouse     8284 6692 0.5531517 0.44684829
#>  Married-spouse-absent   384   34 0.9186603 0.08133971
#>  Never-married         10192  491 0.9540391 0.04596087
#>  Separated               959   66 0.9356098 0.06439024
#>  Widowed                 908   85 0.9144008 0.08559919
#> 
#> Infomation value: 1.338799
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$OCCUPATION
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: OCCUPATION of type: character
#> 
#>                       0    1         0           1
#>  ?                 1652  191 0.8963646 0.103635377
#>  Adm-clerical      3263  507 0.8655172 0.134482759
#>  Armed-Forces         8    1 0.8888889 0.111111111
#>  Craft-repair      3170  929 0.7733594 0.226640644
#>  Exec-managerial   2098 1968 0.5159862 0.484013773
#>  Farming-fishing    879  115 0.8843058 0.115694165
#>  Handlers-cleaners 1284   86 0.9372263 0.062773723
#>  Machine-op-inspct 1752  250 0.8751249 0.124875125
#>  Other-service     3158  137 0.9584219 0.041578149
#>  Priv-house-serv    148    1 0.9932886 0.006711409
#>  Prof-specialty    2281 1859 0.5509662 0.449033816
#>  Protective-serv    438  211 0.6748844 0.325115562
#>  Sales             2667  983 0.7306849 0.269315068
#>  Tech-support       645  283 0.6950431 0.304956897
#>  Transport-moving  1277  320 0.7996243 0.200375704
#> 
#> Infomation value: 0.776134
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$RELATIONSHIP
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: RELATIONSHIP of type: character
#> 
#>                    0    1         0          1
#>  Husband        7275 5918 0.5514288 0.44857121
#>  Not-in-family  7449  856 0.8969296 0.10307044
#>  Other-relative  944   37 0.9622834 0.03771662
#>  Own-child      5001   67 0.9867798 0.01322021
#>  Unmarried      3228  218 0.9367382 0.06326175
#>  Wife            823  745 0.5248724 0.47512755
#> 
#> Infomation value: 1.535553
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$SEX
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: SEX of type: character
#> 
#>             0    1         0         1
#>  Female  9592 1179 0.8905394 0.1094606
#>  Male   15128 6662 0.6942634 0.3057366
#> 
#> Infomation value: 0.303287
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$HOURSPERWEEK
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: HOURSPERWEEK of type: numeric
#> 
#>                       0    1         0          1
#> [a] from 1 to 20   1587  117 0.9313380 0.06866197
#> [b] from 20 to 32  3310  236 0.9334461 0.06655386
#> [c] from 32 to 40  2128  385 0.8467967 0.15320334
#> [d] from 40 to 45 12398 3437 0.7829492 0.21705084
#> [e] from 45 to 50  1615  886 0.6457417 0.35425830
#> [f] from 50 to 56  2060 1670 0.5522788 0.44772118
#> [g] from 56 to 99  1622 1110 0.5937042 0.40629575
#> 
#> Infomation value: 0.462279
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$NATIVECOUNTRY
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: NATIVECOUNTRY of type: character
#> 
#>                                 0    1         0          1
#>  ?                            437  146 0.7495712 0.25042882
#>  Cambodia                      12    7 0.6315789 0.36842105
#>  Canada                        82   39 0.6776860 0.32231405
#>  China                         55   20 0.7333333 0.26666667
#>  Columbia                      57    2 0.9661017 0.03389831
#>  Cuba                          70   25 0.7368421 0.26315789
#>  Dominican-Republic            68    2 0.9714286 0.02857143
#>  Ecuador                       24    4 0.8571429 0.14285714
#>  El-Salvador                   97    9 0.9150943 0.08490566
#>  England                       60   30 0.6666667 0.33333333
#>  France                        17   12 0.5862069 0.41379310
#>  Germany                       93   44 0.6788321 0.32116788
#>  Greece                        21    8 0.7241379 0.27586207
#>  Guatemala                     61    3 0.9531250 0.04687500
#>  Haiti                         40    4 0.9090909 0.09090909
#>  Holand-Netherlands             1    0 1.0000000 0.00000000
#>  Honduras                      12    1 0.9230769 0.07692308
#>  Hong                          14    6 0.7000000 0.30000000
#>  Hungary                       10    3 0.7692308 0.23076923
#>  India                         60   40 0.6000000 0.40000000
#>  Iran                          25   18 0.5813953 0.41860465
#>  Ireland                       19    5 0.7916667 0.20833333
#>  Italy                         48   25 0.6575342 0.34246575
#>  Jamaica                       71   10 0.8765432 0.12345679
#>  Japan                         38   24 0.6129032 0.38709677
#>  Laos                          16    2 0.8888889 0.11111111
#>  Mexico                       610   33 0.9486781 0.05132193
#>  Nicaragua                     32    2 0.9411765 0.05882353
#>  Outlying-US(Guam-USVI-etc)    14    0 1.0000000 0.00000000
#>  Peru                          29    2 0.9354839 0.06451613
#>  Philippines                  137   61 0.6919192 0.30808081
#>  Poland                        48   12 0.8000000 0.20000000
#>  Portugal                      33    4 0.8918919 0.10810811
#>  Puerto-Rico                  102   12 0.8947368 0.10526316
#>  Scotland                       9    3 0.7500000 0.25000000
#>  South                         64   16 0.8000000 0.20000000
#>  Taiwan                        31   20 0.6078431 0.39215686
#>  Thailand                      15    3 0.8333333 0.16666667
#>  Trinadad&Tobago               17    2 0.8947368 0.10526316
#>  United-States              21999 7171 0.7541652 0.24583476
#>  Vietnam                       62    5 0.9253731 0.07462687
#>  Yugoslavia                    10    6 0.6250000 0.37500000
#> 
#> Infomation value: 0.081463
#> 
#> There is a `summary` method available as well as various `plot` function.
#> 
#> $target_reference_level
#> [1] 1
#> 
#> attr(,"class")
#> [1] "targeter" "list"    
focus(t, n = 3)
#> $dataname
#> [1] "adult"
#> 
#> $target
#> [1] "ABOVE50K"
#> 
#> $description_data
#> NULL
#> 
#> $description_target
#> NULL
#> 
#> $target_type
#> [1] "binary"
#> 
#> $target_stats
#>   value count nNA percNA      perc
#> 1     0 24720   0      0 0.7591904
#> 2     1  7841   0      0 0.2408096
#> 
#> $analysis
#> [1] "Analyse"
#> 
#> $date
#> [1] "2023-09-13"
#> 
#> $profiles
#> $profiles$AGE
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: AGE of type: numeric
#> 
#>                      0    1         0            1
#> [a] from 17 to 21 2408    2 0.9991701 0.0008298755
#> [b] from 21 to 24 2334   28 0.9881456 0.0118543607
#> [c] from 24 to 28 3031  228 0.9300399 0.0699601105
#> [d] from 28 to 31 2117  424 0.8331366 0.1668634396
#> [e] from 31 to 34 2028  563 0.7827094 0.2172906214
#> [f] from 34 to 37 1937  723 0.7281955 0.2718045113
#> [g] from 37 to 40 1649  852 0.6593363 0.3406637345
#> [h] from 40 to 44 2062 1090 0.6541878 0.3458121827
#> [i] from 44 to 48 1754 1149 0.6042025 0.3957974509
#> [j] from 48 to 52 1391  926 0.6003453 0.3996547259
#> [k] from 52 to 59 1795 1071 0.6263084 0.3736915562
#> [l] from 59 to 90 2214  785 0.7382461 0.2617539180
#> 
#> Infomation value: 1.218432
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$MARITALSTATUS
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: MARITALSTATUS of type: character
#> 
#>                            0    1         0          1
#>  Divorced               3980  463 0.8957911 0.10420887
#>  Married-AF-spouse        13   10 0.5652174 0.43478261
#>  Married-civ-spouse     8284 6692 0.5531517 0.44684829
#>  Married-spouse-absent   384   34 0.9186603 0.08133971
#>  Never-married         10192  491 0.9540391 0.04596087
#>  Separated               959   66 0.9356098 0.06439024
#>  Widowed                 908   85 0.9144008 0.08559919
#> 
#> Infomation value: 1.338799
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$RELATIONSHIP
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: RELATIONSHIP of type: character
#> 
#>                    0    1         0          1
#>  Husband        7275 5918 0.5514288 0.44857121
#>  Not-in-family  7449  856 0.8969296 0.10307044
#>  Other-relative  944   37 0.9622834 0.03771662
#>  Own-child      5001   67 0.9867798 0.01322021
#>  Unmarried      3228  218 0.9367382 0.06326175
#>  Wife            823  745 0.5248724 0.47512755
#> 
#> Infomation value: 1.535553
#> 
#> There is a `summary` method available as well as various `plot` function.
#> 
#> $target_reference_level
#> [1] 1
#> 
#> attr(,"class")
#> [1] "targeter" "list"    
focus(t, nmin = 500, criteria = "index.max.index")
#> $dataname
#> [1] "adult"
#> 
#> $target
#> [1] "ABOVE50K"
#> 
#> $description_data
#> NULL
#> 
#> $description_target
#> NULL
#> 
#> $target_type
#> [1] "binary"
#> 
#> $target_stats
#>   value count nNA percNA      perc
#> 1     0 24720   0      0 0.7591904
#> 2     1  7841   0      0 0.2408096
#> 
#> $analysis
#> [1] "Analyse"
#> 
#> $date
#> [1] "2023-09-13"
#> 
#> $profiles
#> $profiles$AGE
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: AGE of type: numeric
#> 
#>                      0    1         0            1
#> [a] from 17 to 21 2408    2 0.9991701 0.0008298755
#> [b] from 21 to 24 2334   28 0.9881456 0.0118543607
#> [c] from 24 to 28 3031  228 0.9300399 0.0699601105
#> [d] from 28 to 31 2117  424 0.8331366 0.1668634396
#> [e] from 31 to 34 2028  563 0.7827094 0.2172906214
#> [f] from 34 to 37 1937  723 0.7281955 0.2718045113
#> [g] from 37 to 40 1649  852 0.6593363 0.3406637345
#> [h] from 40 to 44 2062 1090 0.6541878 0.3458121827
#> [i] from 44 to 48 1754 1149 0.6042025 0.3957974509
#> [j] from 48 to 52 1391  926 0.6003453 0.3996547259
#> [k] from 52 to 59 1795 1071 0.6263084 0.3736915562
#> [l] from 59 to 90 2214  785 0.7382461 0.2617539180
#> 
#> Infomation value: 1.218432
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$WORKCLASS
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: WORKCLASS of type: character
#> 
#>                       0    1         0         1
#>  ?                 1645  191 0.8959695 0.1040305
#>  Federal-gov        589  371 0.6135417 0.3864583
#>  Local-gov         1476  617 0.7052078 0.2947922
#>  Never-worked         7    0 1.0000000 0.0000000
#>  Private          17733 4963 0.7813271 0.2186729
#>  Self-emp-inc       494  622 0.4426523 0.5573477
#>  Self-emp-not-inc  1817  724 0.7150728 0.2849272
#>  State-gov          945  353 0.7280431 0.2719569
#>  Without-pay         14    0 1.0000000 0.0000000
#> 
#> Infomation value: 0.166504
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$FNLWGT
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: FNLWGT of type: numeric
#> 
#>                               0   1         0         1
#> [a] from 12285 to 55900    2104 610 0.7752395 0.2247605
#> [b] from 55900 to 98253    2113 600 0.7788426 0.2211574
#> [c] from 98253 to 117827   2010 703 0.7408773 0.2591227
#> [d] from 117827 to 141067  2003 711 0.7380251 0.2619749
#> [e] from 141067 to 162343  1991 721 0.7341445 0.2658555
#> [f] from 162343 to 178356  2021 692 0.7449318 0.2550682
#> [g] from 178356 to 193160  2015 700 0.7421731 0.2578269
#> [h] from 193160 to 210474  2053 657 0.7575646 0.2424354
#> [i] from 210474 to 237051  2119 597 0.7801915 0.2198085
#> [j] from 237051 to 279465  2136 578 0.7870302 0.2129698
#> [k] from 279465 to 342730  2091 621 0.7710177 0.2289823
#> [l] from 342730 to 1484705 2064 651 0.7602210 0.2397790
#> 
#> Infomation value: 0.009781
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$EDUCATIONNUM
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: EDUCATIONNUM of type: numeric
#> 
#>                      0    1         0          1
#> [a] from 1 to 7   2494  151 0.9429112 0.05708885
#> [b] from 7 to 9   1515   93 0.9421642 0.05783582
#> [c] from 9 to 10  8826 1675 0.8404914 0.15950862
#> [d] from 10 to 12 6925 1748 0.7984550 0.20154502
#> [e] from 12 to 13  802  265 0.7516401 0.24835989
#> [f] from 13 to 16 4158 3909 0.5154332 0.48456675
#> 
#> Infomation value: 0.662038
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$MARITALSTATUS
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: MARITALSTATUS of type: character
#> 
#>                            0    1         0          1
#>  Divorced               3980  463 0.8957911 0.10420887
#>  Married-AF-spouse        13   10 0.5652174 0.43478261
#>  Married-civ-spouse     8284 6692 0.5531517 0.44684829
#>  Married-spouse-absent   384   34 0.9186603 0.08133971
#>  Never-married         10192  491 0.9540391 0.04596087
#>  Separated               959   66 0.9356098 0.06439024
#>  Widowed                 908   85 0.9144008 0.08559919
#> 
#> Infomation value: 1.338799
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$OCCUPATION
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: OCCUPATION of type: character
#> 
#>                       0    1         0           1
#>  ?                 1652  191 0.8963646 0.103635377
#>  Adm-clerical      3263  507 0.8655172 0.134482759
#>  Armed-Forces         8    1 0.8888889 0.111111111
#>  Craft-repair      3170  929 0.7733594 0.226640644
#>  Exec-managerial   2098 1968 0.5159862 0.484013773
#>  Farming-fishing    879  115 0.8843058 0.115694165
#>  Handlers-cleaners 1284   86 0.9372263 0.062773723
#>  Machine-op-inspct 1752  250 0.8751249 0.124875125
#>  Other-service     3158  137 0.9584219 0.041578149
#>  Priv-house-serv    148    1 0.9932886 0.006711409
#>  Prof-specialty    2281 1859 0.5509662 0.449033816
#>  Protective-serv    438  211 0.6748844 0.325115562
#>  Sales             2667  983 0.7306849 0.269315068
#>  Tech-support       645  283 0.6950431 0.304956897
#>  Transport-moving  1277  320 0.7996243 0.200375704
#> 
#> Infomation value: 0.776134
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$RELATIONSHIP
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: RELATIONSHIP of type: character
#> 
#>                    0    1         0          1
#>  Husband        7275 5918 0.5514288 0.44857121
#>  Not-in-family  7449  856 0.8969296 0.10307044
#>  Other-relative  944   37 0.9622834 0.03771662
#>  Own-child      5001   67 0.9867798 0.01322021
#>  Unmarried      3228  218 0.9367382 0.06326175
#>  Wife            823  745 0.5248724 0.47512755
#> 
#> Infomation value: 1.535553
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$SEX
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: SEX of type: character
#> 
#>             0    1         0         1
#>  Female  9592 1179 0.8905394 0.1094606
#>  Male   15128 6662 0.6942634 0.3057366
#> 
#> Infomation value: 0.303287
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$CAPITALGAIN
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: CAPITALGAIN of type: numeric
#> 
#>                         0    1         0         1
#> [a] from 0 to 99999 24720 7841 0.7591904 0.2408096
#> 
#> Infomation value: 0
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$HOURSPERWEEK
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: HOURSPERWEEK of type: numeric
#> 
#>                       0    1         0          1
#> [a] from 1 to 20   1587  117 0.9313380 0.06866197
#> [b] from 20 to 32  3310  236 0.9334461 0.06655386
#> [c] from 32 to 40  2128  385 0.8467967 0.15320334
#> [d] from 40 to 45 12398 3437 0.7829492 0.21705084
#> [e] from 45 to 50  1615  886 0.6457417 0.35425830
#> [f] from 50 to 56  2060 1670 0.5522788 0.44772118
#> [g] from 56 to 99  1622 1110 0.5937042 0.40629575
#> 
#> Infomation value: 0.462279
#> 
#> There is a `summary` method available as well as various `plot` function.
#> 
#> $target_reference_level
#> [1] 1
#> 
#> attr(,"class")
#> [1] "targeter" "list"    
focus(t, criteria = "chisquare")
#> $dataname
#> [1] "adult"
#> 
#> $target
#> [1] "ABOVE50K"
#> 
#> $description_data
#> NULL
#> 
#> $description_target
#> NULL
#> 
#> $target_type
#> [1] "binary"
#> 
#> $target_stats
#>   value count nNA percNA      perc
#> 1     0 24720   0      0 0.7591904
#> 2     1  7841   0      0 0.2408096
#> 
#> $analysis
#> [1] "Analyse"
#> 
#> $date
#> [1] "2023-09-13"
#> 
#> $profiles
#> $profiles$AGE
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: AGE of type: numeric
#> 
#>                      0    1         0            1
#> [a] from 17 to 21 2408    2 0.9991701 0.0008298755
#> [b] from 21 to 24 2334   28 0.9881456 0.0118543607
#> [c] from 24 to 28 3031  228 0.9300399 0.0699601105
#> [d] from 28 to 31 2117  424 0.8331366 0.1668634396
#> [e] from 31 to 34 2028  563 0.7827094 0.2172906214
#> [f] from 34 to 37 1937  723 0.7281955 0.2718045113
#> [g] from 37 to 40 1649  852 0.6593363 0.3406637345
#> [h] from 40 to 44 2062 1090 0.6541878 0.3458121827
#> [i] from 44 to 48 1754 1149 0.6042025 0.3957974509
#> [j] from 48 to 52 1391  926 0.6003453 0.3996547259
#> [k] from 52 to 59 1795 1071 0.6263084 0.3736915562
#> [l] from 59 to 90 2214  785 0.7382461 0.2617539180
#> 
#> Infomation value: 1.218432
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$WORKCLASS
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: WORKCLASS of type: character
#> 
#>                       0    1         0         1
#>  ?                 1645  191 0.8959695 0.1040305
#>  Federal-gov        589  371 0.6135417 0.3864583
#>  Local-gov         1476  617 0.7052078 0.2947922
#>  Never-worked         7    0 1.0000000 0.0000000
#>  Private          17733 4963 0.7813271 0.2186729
#>  Self-emp-inc       494  622 0.4426523 0.5573477
#>  Self-emp-not-inc  1817  724 0.7150728 0.2849272
#>  State-gov          945  353 0.7280431 0.2719569
#>  Without-pay         14    0 1.0000000 0.0000000
#> 
#> Infomation value: 0.166504
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$EDUCATION
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: EDUCATION of type: character
#> 
#>                  0    1         0          1
#>  10th          871   62 0.9335477 0.06645230
#>  11th         1115   60 0.9489362 0.05106383
#>  12th          400   33 0.9237875 0.07621247
#>  1st-4th       162    6 0.9642857 0.03571429
#>  5th-6th       317   16 0.9519520 0.04804805
#>  7th-8th       606   40 0.9380805 0.06191950
#>  9th           487   27 0.9474708 0.05252918
#>  Assoc-acdm    802  265 0.7516401 0.24835989
#>  Assoc-voc    1021  361 0.7387844 0.26121563
#>  Bachelors    3134 2221 0.5852474 0.41475257
#>  Doctorate     107  306 0.2590799 0.74092010
#>  HS-grad      8826 1675 0.8404914 0.15950862
#>  Masters       764  959 0.4434127 0.55658735
#>  Preschool      51    0 1.0000000 0.00000000
#>  Prof-school   153  423 0.2656250 0.73437500
#>  Some-college 5904 1387 0.8097655 0.19023454
#> 
#> Infomation value: 0.753711
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$EDUCATIONNUM
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: EDUCATIONNUM of type: numeric
#> 
#>                      0    1         0          1
#> [a] from 1 to 7   2494  151 0.9429112 0.05708885
#> [b] from 7 to 9   1515   93 0.9421642 0.05783582
#> [c] from 9 to 10  8826 1675 0.8404914 0.15950862
#> [d] from 10 to 12 6925 1748 0.7984550 0.20154502
#> [e] from 12 to 13  802  265 0.7516401 0.24835989
#> [f] from 13 to 16 4158 3909 0.5154332 0.48456675
#> 
#> Infomation value: 0.662038
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$MARITALSTATUS
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: MARITALSTATUS of type: character
#> 
#>                            0    1         0          1
#>  Divorced               3980  463 0.8957911 0.10420887
#>  Married-AF-spouse        13   10 0.5652174 0.43478261
#>  Married-civ-spouse     8284 6692 0.5531517 0.44684829
#>  Married-spouse-absent   384   34 0.9186603 0.08133971
#>  Never-married         10192  491 0.9540391 0.04596087
#>  Separated               959   66 0.9356098 0.06439024
#>  Widowed                 908   85 0.9144008 0.08559919
#> 
#> Infomation value: 1.338799
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$OCCUPATION
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: OCCUPATION of type: character
#> 
#>                       0    1         0           1
#>  ?                 1652  191 0.8963646 0.103635377
#>  Adm-clerical      3263  507 0.8655172 0.134482759
#>  Armed-Forces         8    1 0.8888889 0.111111111
#>  Craft-repair      3170  929 0.7733594 0.226640644
#>  Exec-managerial   2098 1968 0.5159862 0.484013773
#>  Farming-fishing    879  115 0.8843058 0.115694165
#>  Handlers-cleaners 1284   86 0.9372263 0.062773723
#>  Machine-op-inspct 1752  250 0.8751249 0.124875125
#>  Other-service     3158  137 0.9584219 0.041578149
#>  Priv-house-serv    148    1 0.9932886 0.006711409
#>  Prof-specialty    2281 1859 0.5509662 0.449033816
#>  Protective-serv    438  211 0.6748844 0.325115562
#>  Sales             2667  983 0.7306849 0.269315068
#>  Tech-support       645  283 0.6950431 0.304956897
#>  Transport-moving  1277  320 0.7996243 0.200375704
#> 
#> Infomation value: 0.776134
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$RELATIONSHIP
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: RELATIONSHIP of type: character
#> 
#>                    0    1         0          1
#>  Husband        7275 5918 0.5514288 0.44857121
#>  Not-in-family  7449  856 0.8969296 0.10307044
#>  Other-relative  944   37 0.9622834 0.03771662
#>  Own-child      5001   67 0.9867798 0.01322021
#>  Unmarried      3228  218 0.9367382 0.06326175
#>  Wife            823  745 0.5248724 0.47512755
#> 
#> Infomation value: 1.535553
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$RACE
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: RACE of type: character
#> 
#>                         0    1         0          1
#>  Amer-Indian-Eskimo   275   36 0.8842444 0.11575563
#>  Asian-Pac-Islander   763  276 0.7343600 0.26564004
#>  Black               2737  387 0.8761204 0.12387964
#>  Other                246   25 0.9077491 0.09225092
#>  White              20699 7117 0.7441401 0.25585994
#> 
#> Infomation value: 0.069292
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$SEX
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: SEX of type: character
#> 
#>             0    1         0         1
#>  Female  9592 1179 0.8905394 0.1094606
#>  Male   15128 6662 0.6942634 0.3057366
#> 
#> Infomation value: 0.303287
#> 
#> There is a `summary` method available as well as various `plot` function.
#> $profiles$HOURSPERWEEK
#> 
#> Targeter crossvar profiling object with following properties:
#> 	Target: ABOVE50K  of type: binary  (target level:1 )
#> 	Crossed with variable: HOURSPERWEEK of type: numeric
#> 
#>                       0    1         0          1
#> [a] from 1 to 20   1587  117 0.9313380 0.06866197
#> [b] from 20 to 32  3310  236 0.9334461 0.06655386
#> [c] from 32 to 40  2128  385 0.8467967 0.15320334
#> [d] from 40 to 45 12398 3437 0.7829492 0.21705084
#> [e] from 45 to 50  1615  886 0.6457417 0.35425830
#> [f] from 50 to 56  2060 1670 0.5522788 0.44772118
#> [g] from 56 to 99  1622 1110 0.5937042 0.40629575
#> 
#> Infomation value: 0.462279
#> 
#> There is a `summary` method available as well as various `plot` function.
#> 
#> $target_reference_level
#> [1] 1
#> 
#> attr(,"class")
#> [1] "targeter" "list"