Plot a targeter tartree object

This function creates a plot for a targeter decision tree model build with `tartree` function If format is static, one will use rpart.plot package. If interactive, one will use visTree.

Usage

# S3 method for class 'tartree'
plot(
  x,
  format = "static",
  prefix = "target = ",
  type = 2,
  yesno = 2,
  branch = 0,
  branch.type = 5,
  box.palette = "Blues",
  shadow.col = 0,
  cex = 0.7,
  left = FALSE,
  ...
)

Arguments

x

- a decision tree as fitted by `tartree` function

format

see `rpart.plot` help

prefix

-see `rpart.plot` help

type

One of 'static' (use rpart.plot) or interactive (use visTree) Parameter expansion is used so that one can use "i". Default: static.

yesno

-see `rpart.plot` help

branch

-see `rpart.plot` help

branch.type

-see `rpart.plot` help

box.palette

-see `rpart.plot` help

shadow.col

-see `rpart.plot` help

cex

-see `rpart.plot` help

left

-see `rpart.plot` help

...

- eventuial additional parameters to be passed to `rpart.plot` or visTree