report

This function creates an automatic report according to a predefined template in the package or a user-generated template.

Usage

report(
  object,
  template = NULL,
  summary_object = NULL,
  browse = FALSE,
  ntop = NULL,
  nmin = 20,
  criteria = c("IV"),
  min_criteria = NULL,
  metadata = NULL,
  force_vars = character(),
  output_format = c("html", "pdf", "word"),
  output_file = NULL,
  output_dir = tempdir(),
  ...
)

Arguments

object

an object of calss "targeter".

template

template file path. By default, it put to NULL and take the default template.

summary_object

an object of class "summary.targeter": pre-computed summary of object.

browse

by default TRUE : print in the browse the output.

ntop

integer - number of variables to show in the report. For more information, see the function focus By default, NULL i.e. all variables are taken.

nmin

integer - minimum number of profiles in the selected class. For more information, see the function focus.

criteria

character - By default, the value is index.max.index. For more information, see the function focus.

min_criteria

- By default NULL. If it's specified, only the observations which have criteria >= min_criteria.

metadata

data.frame - if metadata is loaded in R environment, label of the variables can be used. Default value (NULL) corresponds to no metadata available. The label will be used for the title and the x-axis of the graph.

force_vars

- character, list of variables that will be kept and then won't never be filtered.

output_format

- By default it's html. For a pdf, output_format="pdf_document" and for a word, output_format="word_document".

output_file

- name of the output file to be generated. If NULL (default) an, automated name will be generated.

output_dir

- output directory, default: tempdir().

...

other parameter from the function render.

Value

path to the generated report. Side effect would be to open a document, if browse is TRUE.

See also

Examples

if (FALSE) {
t <- targeter(adult,target ="ABOVE50K",analysis_name="Analyse",naming_conventions=FALSE)
report(t,output_format="pdf_document")
report(ntop=5)
}