report

This function creates an automatic report according to a predefined template in the package or a user-generated template. Create a report from a targeter report object

Usage

report(
  object,
  summary_object = NULL,
  metadata = NULL,
  format = "html",
  nmax = 100,
  template = NULL,
  quarto_root_dir = ".",
  quarto_targeters_project_dir = "targeter-reports",
  quarto_project_template = getOption("targeter.template",
    file.path(find.package("targeter", lib.loc = .libPaths()), "ressources",
    "_targeter_project_default.yml")),
  quarto_project_brandfile = getOption("targeter.brandfile",
    file.path(find.package("targeter", lib.loc = .libPaths()), "ressources",
    "_targeter_brand_default.yml")),
  targeter_sub_folder = paste0("targeter-report-", format(Sys.time(), format =
    "%Y-%m-%d_%H%M%S")),
  pptx_reference_doc = NULL,
  output_file = "index",
  title = object$analysis,
  author = getOption("targeter.author", "welovedatascience targeter"),
  fullplot_which_plot = "1:2",
  show_tables = FALSE,
  debug = FALSE,
  render = TRUE,
  logo = NULL,
  freeze = TRUE,
  custom_fields = list(report_type = "targeter", freeze = freeze),
  ...
)

Arguments

object

- an object of class targeter

summary_object

an object of class "summary.targeter": pre-computed summary of object.

metadata

- data.frame - if metadata is loaded in R environment, label of the variables can be used. Default value (NULL) corresponds to no metadata available. The label will be used for the title and the x-axis of the graph.

format

- Target output format (defaults to "html"). The option "all" will render all formats.

nmax

- integer: maximum number of variables to be used. Default 100. use NULL to force all variables to be present in report.See notes.

template

- character. path to a Quarto qmd document that will be used to generate presentation. If NULL (default), we will use targeter package default template.

quarto_root_dir

(default "./targeter-reports") Path to a root folder where quarto project folder will be avaible. Default to "." so that calling targeter first time in an analyses project will initiate everything for user.

quarto_targeters_project_dir

(default "targeter-reports", within quarto_root_dir) Path to a folder where all targeter reports will be stored. Using a pre-existing projetc allows benfiting from its settings. (you might want to put brand_file to empty or use your own brand file)

quarto_project_template

- character. Path to a default Quarto template that will be used as Quarto reference-doc for all formats. If NULL (default), we will use targeter package default templates.

quarto_project_brandfile

- character. Path to a default Quarto brandfile that will be used as Quarto reference-doc for all formats. If NULL (default), we will use targeter package default brandfile.

targeter_sub_folder

- character. Output directory - default to create a new dedicated directory in working folder.

pptx_reference_doc

- character. Path to a default PPTX template that will be used as Quarto reference-doc for pppx formzt If NULL (default), we will use targeter package default templates for PPTX. If empty string "", we won't use any template.

output_file

- character, name of output file (without file extension). default value: index

title

- character, title for the presentation. Overide default title that consists in taking title for targeter object 'analysis' slot.

author

- character: presentatiopn author.

fullplot_which_plot

- character, selection of plots (cf `fullplot` documentation)

show_tables

- logical whetheer to print tables in seprate slides in addition of graphics. Might ot produce nice outputs as tables are usually large.

debug

- logical. If TRUE, the function will print debug information (passed to Quarto call).

render

- logical. If TRUE (default), the report will be rendered and saved in the target_sub_folder. If FALSE, the report will be generated but not rendered. Useful for debugging.

logo

- character: path to a logo file. If NULL (default), we will take package welovedatascience logo. If empty, no logo will be used. Only used by revealjs format.

freeze

- logical: default TRUE. If TRUE, the report will be frozen (no more changes can be made). If FALSE, the report will be editable. Frozen reports renders can always be bypassed by explicit call to quarto.

custom_fields

- list of custom fields to be added to the report.

...

additional parameters to be passed to quarto, for instance allowing to pass/overwite any YAML set-up

Value

invisibly returns path to the generated specific file (unique format) or folder (several formats)

See also

Examples

if (FALSE) { # \dontrun{
tar <- targeter(adult, target ="ABOVE50K", analysis_name="Analyse",
naming_conventions=FALSE)
report(tar,  author = "this is me")
report(tar,  format= c("pptx","pdf","docx"),author = "this is me")
} # }