Create a presentation from a targeter report object

This function generates a presentation (slides), currently targeting powerpoint (other formats to come).

Usage

slidify(
  object,
  metadata = NULL,
  format = "pptx",
  template = NULL,
  pptx_reference_doc = NULL,
  revealjs_template = "",
  output_dir = ".",
  output_file = NULL,
  delete_temporary_files = TRUE,
  title = object$analysis,
  author = getOption("targeter.author", "welovedatascience targeter"),
  fullplot_which_plot = "1:2",
  tables = FALSE,
  logo = NULL,
  ...
)

Arguments

object

- an object of class targeter

metadata

- not used currently

format

- character, one of "pptx", "revealjs"

template

- character. path to a Quarto qmd document that will be used to generate presentation. If NULL (default), we will use targeter package default template.

pptx_reference_doc

- character. Path to a default PPTX template that will be used as Q"uarto reference-doc If NULL (default), we will use targeter package default templates for PPTX. If empty string "", we won't use any template.

revealjs_template

- character. Path to a default revealjs template If empty string (default), no revealjs template will be used.

output_dir

- character. Output directory - default to working directory

output_file

- character, name of output file (with extension). if NULL (default), slidify will generate a temporary name.

delete_temporary_files

- logical. Whether we delete temporary files or not once presentation is generated. Default: TRUE. On might want to use FALSE for debugging purpose.

title

- character, title for the presentation. Overide default title that consists in taking title for targeter object 'analysis' slot.

author

- character: presentatiopn author.

fullplot_which_plot

- character, selection of plots (cf `fullplot` documentation)

tables

- logical whetheer to print tables in seprate slides in addition of graphics. Might ot produce nice outputs as tables are usually large.

logo

- character: path to a logo file. If NULL (default), we will take package welovedatascience logo. If empty, no logo will be used. Only used by revealjs format.

...

additional parameters to be passed to quarto, for instance allowing to pass/overwite any YAML set-up

Value

inviisibly returns path to the generated presentation

Examples

if (FALSE) { # \dontrun{
tar <- targeter(adult, target ="ABOVE50K", analysis_name="Analyse",
naming_conventions=FALSE)
slidify(tar, format = "pptx", author = "this is me")
} # }