summary.targeter

summary method for targeter object

Usage

# S3 method for class 'targeter'
summary(
  object,
  extra_stats = FALSE,
  criteria = c("IV", "index.max.index", "chisquare", "pvalue", "index.max.count",
    "index.max.props"),
  ...
)

Arguments

object

an object of class targeter

extra_stats

boolean - whether function returns some additional associations statistics (default: FALSE)

criteria

which criteria is used to sort the summary table (default: IV)

...

additional parameters (currently not used)

Value

a data.table summary table

Details

summary method invoked on a targeter object loops over all individual profiles/crossvar/associations and derive statistics (cf summary.crossvar). Returned table contains thus one row per explanatory variable analyzed in targeter and columns for associations measures / metrics.

See also

Examples

data(adult)
t <- targeter(adult,target ="ABOVE50K",analysis_name="Analyse")
#> Error in data.table(variable = sort(unique(c(select_vars, exclude_vars,     names(data))))): could not find function "data.table"
summary(t)
#> Error in object[[i]]: object of type 'closure' is not subsettable
summary(t, extra_stats = TRUE)
#> Error in object[[i]]: object of type 'closure' is not subsettable