table_crossvar: build a meaningful table from a crossvar object.

This functions builds a dataframe that gather most relevant information from a crossvar object. Useful in reports, associated with kableExtra or other tables output generating engine.

Usage

table_crossvar(x, round_digits = 3, counts_and = c("props", "target%"))

Arguments

x

crossvar - crossvar object (stored in targeter$profiles slot).

round_digits

integer - number of digits to rounds some values.

counts_and

- character - output will always contain counts, we can add row percentages and column percentage for target (both always by default)

Value

a data.frame

Examples

tar <- targeter(adult, target = 'ABOVE50K')
#> Error in data.table(variable = sort(unique(c(select_vars, exclude_vars,     names(data))))): could not find function "data.table"
table_crossvar(tar$profiles[['EDUCATION']])
#> Error in tar$profiles: object of type 'closure' is not subsettable